add deviceID to schema
