+++ docs/PySpark/structured-streaming-pyspark.md
@@ -246,7 +246,9 @@ Each row in the source has the following schema:
 | sequenceNumber | long |
 | enqueuedTime | timestamp |
 | publisher | string |
-| partitionKey | string | 
+| partitionKey | string |
+| properties | map[string, json] |
+| connectionDeviceID | string |
 
 ## Writing Data to Event Hubs
